Remove deprecated files and configurations, including .prettierignore, .prettierrc, and app-related files, to streamline the project structure and focus on core functionality.
