+++ .gitignore
@@ -3,8 +3,3 @@ node_modules
 dist
 .DS_Store
 coverage
-
-eslint.config.js
-
-app/LICENSE.txt
-app/README.md
