--- .prettierignore
@@ -1,10 +1,0 @@
-.cache
-package.json
-package-lock.json
-public
-
-node_modules
-build
-dist
-res
-coverage
