--- .prettierrc
@@ -1,15 +1,0 @@
-{
-  "ln": 1,
-  "endOfLine": "lf",
-  "printWidth": 130,
-  "proseWrap": "preserve",
-  "semi": true,
-  "singleQuote": true,
-  "useTabs": false,
-  "tabWidth": 2,
-  "arrowParens": "always",
-  "trailingComma": "es5",
-  "bracketSpacing": true,
-  "jsxBracketSameLine": false,
-  "jsxSingleQuote": true
-}
