--- app/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "jest-test-clean-db",
-  "version": "1.1.13",
-  "description": "Clean DB, with the migration files pre-applied, per Jest test. No boilerplate code!",
-  "main": "dist/index.js",
-  "files": [
-    "dist/**/*",
-    "globalSetup.js",
-    "globalTeardown.js",
-    "../LICENSE.txt",
-    "../README.md"
-  ],
-  "scripts": {
-    "task:release": "dotenv -e .env np -- --no-2fa --no-cleanup",
-    "task:test_OLD": "jest --runInBand --detectOpenHandles --forceExit",
-    "task:typescript": "tsc --pretty",
-    "task:typescript_elaborated": "tsc -sourcemap -incremental --pretty",
-    "release": "pnpm run task:typescript && pnpm run task:release",
-    "build": "pnpm run task:typescript",
-    "start": "ts-node -r source-map-support/register -r dotenv/config index.ts",
-    "test": "echo Leave this line, as it is required by the np package",
-    "preinstall": "npx only-allow pnpm"
-  },
-  "dependencies": {
-    "jest": "^29.7.0",
-    "pg": "^8.3.3",
-    "typeorm": "^0.3.17"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.23",
-    "@types/node": "^14.6.4",
-    "dotenv": "^16.0.1",
-    "jest-silent-reporter": "^0.5.0",
-    "np": "^10.0.7",
-    "ts-jest": "^29.3.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
-  },
-  "author": "Asaf Shifer",
-  "license": "MIT"
-}
