--- demo/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "jest_postgresql_demo",
-  "type": "module",
-  "description": "",
-  "scripts": {
-    "test_OLD": "jest --runInBand --detectOpenHandles --forceExit",
-    "test": "jest --detectOpenHandles --forceExit"
-  },
-  "dependencies": {
-    "pg": "^8.3.3",
-    "typeorm": "^0.3.17",
-    "typeorm-extension": "^2.8.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.23",
-    "@types/node": "^14.6.4",
-    "dotenv": "^16.0.1",
-    "jest": "^29.7.0",
-    "jest-silent-reporter": "^0.5.0",
-    "jest-test-clean-db": "workspace:*",
-    "ts-jest": "^29.3.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
-  },
-  "author": "",
-  "license": "ISC"
-}
