--- demo/tsconfig.json
@@ -1,31 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true,
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "target": "ES2017",
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noEmitOnError": true,
-    "allowJs": false,
-    "alwaysStrict": true,
-    "strictBindCallApply": true,
-    "strictFunctionTypes": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "emitDecoratorMetadata": true,
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "outDir": "dist",
-    "declaration": true,
-    "baseUrl": "app",
-    // "typeRoots": ["./node_modules/**/*"],
-    "paths": {
-      "*": ["node_modules/*"]
-    }
-  },
-  // "include": ["./*"],
-}
