+++ globalTearDown.ts
@@ -1,0 +1,10 @@
+import { getMainDataSource } from './mainDataSource_singleton';
+import { dumpSharedGlobalData } from './sharedGlobalData';
+import { consoleDebug } from './utils';
+
+export default async () => {
+    const mainDataSource = getMainDataSource();
+    consoleDebug(`globalTeardown - wasEverInitialized: ${mainDataSource.wasEverInitialized()}`);
+    await mainDataSource.closeAndDelete_templateDb();
+    return dumpSharedGlobalData();
+  };
