+++ package.json
@@ -1,29 +1,43 @@
 {
-  "name": "root",
-  "description": "Clean DB, with the migration files pre-applied, per Jest test. No boilerplate code!",
-  "main": "index.js",
-  "scripts": {
-    "release": "pnpm install && pnpm run task:copy_meta_files && pnpm --filter ./app run release && pnpm run task:remove_meta_files",
-    "task:test": "pnpm --parallel run test",
-    "task:copy_meta_files": "shx cp ./{LICENSE.txt,README.md} ./app",
-    "task:remove_meta_files": "shx rm ./app/LICENSE.txt ./app/README.md",
-    "build": "pnpm --parallel run task:typescript",
-    "test": "pnpm run build && pnpm run task:test",
-    "lint": "tsc --target esnext --module NodeNext --moduleResolution NodeNext eslint.config.ts && eslint . --fix",
-    "preinstall": "npx only-allow pnpm"
-  },
-  "devDependencies": {
-    "@types/eslint__js": "^8.42.3",
-    "@types/node": "^14.18.51",
-    "@typescript-eslint/eslint-plugin": "^8.0.1",
-    "@typescript-eslint/parser": "^8.0.1",
-    "eslint": "^9.8.0",
-    "eslint-plugin-jest": "^28.8.0",
-    "globals": "^15.9.0",
-    "shx": "^0.3.4",
-    "typescript": "^4.7.4",
-    "typescript-eslint": "^8.0.1"
-  },
-  "author": "Asaf Shifer",
-  "license": "MIT"
-}
+    "name": "jest-test-clean-db-postgres",
+    "version": "0.1.11",
+    "description": "A utility package for testing TypeORM applications with Jest and PostgreSQL database. This package helps manage database cleanup and initialization during test runs.",
+    "main": "dist/index.js",
+    "files": [
+      "dist/**/*",
+      "./globalSetup.js",
+      "./globalTeardown.js"
+    ],
+    "scripts": {
+      "task:release": "dotenv -e .env np -- --no-2fa",
+      "task:test_OLD": "jest --runInBand --detectOpenHandles --forceExit",
+      "task:typescript": "tsc --pretty",
+      "task:typescript_elaborated": "tsc -sourcemap -incremental --pretty",
+      "release": "npm run task:typescript && npm run task:release",
+      "build": "npm run task:typescript",
+      "start": "ts-node -r source-map-support/register -r dotenv/config index.ts",
+      "test": "echo Leave this line, as it is required by the np package"
+    },
+    "dependencies": {
+      "pg": "^8.3.3",
+      "typeorm": "^0.3.17"
+    },
+    "devDependencies": {
+      "@types/eslint__js": "^8.42.3",
+      "@types/node": "^14.6.4",
+      "@typescript-eslint/eslint-plugin": "^8.0.1",
+      "@typescript-eslint/parser": "^8.0.1",
+      "dotenv": "^16.0.1",
+      "np": "^10.0.7",
+      "jest-silent-reporter": "^0.5.0",
+      "@types/jest": "^26.0.23",
+      "ts-jest": "^26.4.4",
+      "ts-node": "^10.9.1",
+      "typescript": "^4.7.4",
+      "typescript-eslint": "^8.0.1",
+      "globals": "^15.9.0"
+    },
+    "author": "Arun Wilson",
+    "license": "MIT"
+  }
+
